+++ app/src/main/kotlin/com/maltaisn/notes/receiver/AlarmReceiver.kt
@@ -100,7 +100,7 @@ class AlarmReceiver : BroadcastReceiver() {
             val postponeIntent = Intent(context, NotificationActivity::class.java).apply {
                 action = NotificationActivity.INTENT_ACTION_POSTPONE
                 putExtra(EXTRA_NOTE_ID, noteId)
-                setFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
+                flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
             }
             builder.addAction(R.drawable.ic_calendar,
                 context.getString(R.string.action_postpone),